Add tooltips to configuration fields
Added tooltips for configuration fields such as `Project` and `Connection Setup`. The jenkins widgets no longer have the url-endpoint `https://builds.apache.org/` set as default because if a default url-endpoint is set in the jenkins.js file, the `Setup Configuration` field will not be displayed in the edit view.
Add a CHANGELOG.